# Move some returns in iter to outer layer