Reducing numbers posted on https://github.com/rust-lang/rfcs/pull/1303#issuecomment-148907734 😊

r? @eddyb