@eddyb 2015-10-17
Does this affect any benchmarks?