@bluss 2015-10-17
If I understood @dotdash correctly, there are issues with codegen for matches on tuples like this. Which means that nested `if` may give better code.