@nagisa 2015-10-17
As far as I can tell,

1. We do not have any benchmarks for these functions;
2. This is stylistic change, therefore I deem performance to be out-of-scope here; either way I can’t see these changes changing the performance much as the optimised IRs for outer-return-match and match-inner-return are α-equivalent.