@nagisa 2015-10-18
```rust
test iter::bench_cmp_100000_if                             ... bench:     172,044 ns/iter (+/- 82,675)
test iter::bench_partial_cmp_100000_outer                  ... bench:     200,035 ns/iter (+/- 4,070)
test iter::bench_cmp_100000_inner                          ... bench:     195,487 ns/iter (+/- 9,078)
test iter::bench_partial_cmp_100000_inner                  ... bench:     197,970 ns/iter (+/- 4,188)
```

I made some benchmarks. `_if` is the version using `if`s instead of matches, `_outer` is outer returns and `_inner` is inner returns.

`_inner` and `_outer` are comparable. `_if` seems faster, but with a bigger variance. Variance could be coming from something about my (lack of) bench setup.