@nagisa 2015-10-18
More benchmarks:

### New (if)

```
test iter::bench_cmp_10000                                 ... bench:      17,019 ns/iter (+/- 1,308)
test iter::bench_eq_10000                                  ... bench:      13,443 ns/iter (+/- 325)
test iter::bench_ge_10000                                  ... bench:      18,403 ns/iter (+/- 810)
test iter::bench_gt_10000                                  ... bench:      17,704 ns/iter (+/- 767)
test iter::bench_le_10000                                  ... bench:      20,059 ns/iter (+/- 1,075)
test iter::bench_lt_10000                                  ... bench:      17,957 ns/iter (+/- 605)
test iter::bench_ne_10000                                  ... bench:      13,721 ns/iter (+/- 6,686)
test iter::bench_partial_cmp_10000                         ... bench:      17,179 ns/iter (+/- 5,893)
```

### Old (match-inner-return)

```
test iter::bench_cmp_10000                                 ... bench:      20,229 ns/iter (+/- 957)
test iter::bench_eq_10000                                  ... bench:      16,274 ns/iter (+/- 586)
test iter::bench_ge_10000                                  ... bench:      23,305 ns/iter (+/- 1,312)
test iter::bench_gt_10000                                  ... bench:      22,488 ns/iter (+/- 1,943)
test iter::bench_le_10000                                  ... bench:      22,289 ns/iter (+/- 1,774)
test iter::bench_lt_10000                                  ... bench:      22,656 ns/iter (+/- 2,375)
test iter::bench_ne_10000                                  ... bench:      17,591 ns/iter (+/- 1,379)
test iter::bench_partial_cmp_10000                         ... bench:      21,042 ns/iter (+/- 1,620)
```