@alexcrichton 2015-10-18
Interesting find! Improving benchmarks is always nice, but is there a way to avoid the `unwrap` here? (which is less idiomatic than a `match`)

We may also want to investigate just improving the codegen here so everyone can benefit instead of just us.